Add files via upload
These files will have to remain as PNG files if we want the backgrounds to be transparent. I have tried to make the backgrounds transparent as a JPG file so that it will match the file type of all of the other files in the Image file, but so far, I haven't been able to find a way to do so.